requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic~=1.10
pydantic-settings
sqlalchemy>=2.0
pgvector
psycopg2-binary
redis
rq
boto3
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
python-json-logger
httpx
requests
openai
langchain
python-dotenv
tenacity
pytest
pytest-asyncio
matplotlib
